% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnegie_colors.R
\docType{data}
\name{carnegie_palettes}
\alias{carnegie_palettes}
\title{List of color palettes.}
\format{
An object of class \code{list} of length 3.
}
\usage{
carnegie_palettes
}
\description{
Use \code{\link{render_carnegie_colors}} to construct palettes of desired length.
}
\keyword{datasets}
