% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnegie_colors.R
\name{render_carnegie_colors}
\alias{render_carnegie_colors}
\title{Carnegie colors palette rendering}
\usage{
render_carnegie_colors(name, n)
}
\arguments{
\item{name}{Name of color palette. Choices are:
\code{classic}, \code{secondary}, \code{campus}.}

\item{n}{Number of colors desired. Most palettes have 4-6 colors. Most color
schemes are derived from \href{https://www.cmu.edu/brand/brand-guidelines/visual-identity/colors.html}.
If omitted, uses all colors.}
}
\value{
A vector of colors.
}
\description{
Carnegie colors palette rendering
}
\examples{
render_carnegie_colors("classic")
render_carnegie_colors("secondary", 5)
}
\keyword{colors}
